Assets added, not working right.
Forgot that a coroutine needs to end. Need to add it elsewhere. Not
getting the collisions to work. Been having trouble sending this commit.
Errors been throwing out without the gitignore file.  add new package.
Call me if any developments at 620-228-3041 because I need to go to
work. I'll have tonight and tomorrow to work on this till its done.